move npm installation to the package extraction step
